skill-runner/pyproject.toml
[project]
name = "skill-runner"
version = "0.1.0"
description = "Execute Claude skills from CLI or MCP"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "claude-code-sdk>=0.0.25",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "typer>=0.12.0",
    "rich>=13.0",
]

[project.optional-dependencies]
mcp = ["mcp>=1.0.0"]
dev = ["pytest>=8.0", "ruff>=0.5.0"]

[project.scripts]
skill = "skill_runner.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/skill_runner"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
